version 2.1.3 for pwg 2.1.x
git-svn-id: http://piwigo.org/svn/extensions/music_player@6877 68402e56-0260-453c-a942-63ccdbb3a9ee

+++ maintain.inc.php
@@ -6,7 +6,7 @@ function plugin_install()
 
     
 	    $q = 'CREATE TABLE ' . $prefixeTable . 'mp_playlist (
-id TINYINT NOT NULL AUTO_INCREMENT ,
+id SMALLINT NOT NULL AUTO_INCREMENT ,
 type VARCHAR( 50 ) NOT NULL ,
 url VARCHAR( 255 ) NOT NULL ,
 texte TINYTEXT NOT NULL ,
